network2/requirements.txt
cffi==1.16.0
cryptography==41.0.5
pika==1.3.1
psycopg2-binary==2.9.9
pycparser==2.21
uvloop==0.19.0